backend/requirements.txt
Django==2.2.4
djangorestframework==3.10.2
psycopg2==2.8.4
django-rest-swagger==2.2.0

django-allauth          
django-cors-headers     
django-rest-auth        
django-rest-framework   
djangorestframework     
djangorestframework-jwt 